lims_dev/requirements.txt
requests
flask-lambda-python36
werkzeug==2.0.3
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Cors
Flask-Mail
PyMySQL
